requirements.txt
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
DateTime==5.5
fluent-discourse==1.0.1
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
gspread==6.0.2
httplib2==0.22.0
idna==3.6
joblib==1.4.0
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
StrEnum==0.4.15
tqdm==4.66.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
zope.interface==6.2
